requirements.txt
openai
streamlit
python-docx
python-dotenv
